[Improved] unittesting for test coverage (#24)
