--- tests/main.rs
@@ -1,3 +1,0 @@
-#[test]
-fn it_works() {
-}
